+++ test/term_test.dart
@@ -23,7 +23,7 @@ void tests() {
     });
     test('term interval', () {
       var term = Term.parse('Jan21-Mar21', location);
-      var months = term.interval.splitLeft((dt) => Month.contains(dt));
+      var months = term.interval.splitLeft((dt) => Month.containing(dt));
       expect(months.length, 3);
     });
     test('term is*', () {